Move cleanup job PO<->domain conversion onto the PO class
Per review, relocate toPO/fromPO (and the fileIOProperties JSON helpers)
from IcebergCleanupJobStore into IcebergCleanupJobPO as
fromCleanupJob(job, id, now) and toCleanupJob(), mirroring core's PO
conversion pattern. The store now delegates to these.
